Move tabs to file
